requirements.txt
torch
torchvision
Pillow
numpy
matplotlib
requests
opencv-python